+++ azure_monitor/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Remove request metrics from auto-collection
+  ([#124](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/124))
+
 ## 0.5b.0
 Released 2020-09-24